refactor: seperate `PageEvent` into two classes
there is `PageEvent` and `ArtworksEvent`, `PageEvent` behavior is same as
the previous `PageEvent.root`, and `ArtworksEvent` is equivalent to the
previous `PageEvent.artworks`

`ArtworksEvent` is used for detecting page event on artworks page
