+++ src/Pixiv Downloader/main.ts
@@ -4,18 +4,19 @@ import "./style.css";
 import { waitForElement } from "@owowed/oxi";
 import { requireNonNull } from "@shared/util.ts";
 import { createPbdDownloadManager, updatePbdDownloadManager } from "./illust";
-import { PageEvent } from "./pixiv";
+import { ArtworksEvent } from "./page";
 
-const pageEvent = new PageEvent;
+const artworksEvent = new ArtworksEvent;
+
 let pbdDownloadManager: HTMLDivElement;
 
-pageEvent.artworks.addEventListener("navigate-begin", async () => {
+artworksEvent.addEventListener("navigate-begin", async () => {
     pbdDownloadManager ??= await createPbdDownloadManager();
     const illustDesc = await waitForElement<HTMLDivElement>("figcaption:has(h1):has(footer) div:has(> footer)").then(requireNonNull);
     
     illustDesc.append(pbdDownloadManager);
 });
 
-pageEvent.artworks.addEventListener("navigate", () => {
+artworksEvent.addEventListener("navigate", () => {
     updatePbdDownloadManager(pbdDownloadManager);
 });
